fix problems with AIX/xlC
git-svn-id: svn://svn.code.sf.net/p/cxxtools/code/trunk@1257 15ae5fad-cc11-0410-8fac-bce609e504b0

+++ cxxtools/test/serializer-bench.cpp
@@ -129,7 +129,7 @@ int main(int argc, char* argv[])
             {
                 obj.intValue = n;
                 obj.stringValue = cxxtools::convert<std::string>(n);
-                obj.doubleValue = sqrt(n);
+                obj.doubleValue = sqrt(static_cast<double>(n));
                 v.push_back(obj);
             }
